fix: change api call location
